requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asttokens==3.0.0
blinker==1.9.0
certifi==2025.1.31
click==8.1.8
decorator==5.2.1
distro==1.9.0
executing==2.2.0
Flask==3.1.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
ipython==8.32.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
openai==1.64.0
parso==0.8.4
pexpect==4.9.0
prompt_toolkit==3.0.50
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
python-dotenv==1.0.1
sniffio==1.3.1
stack-data==0.6.3
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
wcwidth==0.2.13
Werkzeug==3.1.3
